requirements.txt
# Streamlit UI
streamlit>=1.28.0

# Vector Database
chromadb>=0.4.0

# Google Gemini API
google-generativeai>=0.3.0

# Environment Variables
python-dotenv>=1.0.0

# Image Processing
Pillow>=10.0.0

# Google Drive API
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Retry Logic
tenacity>=8.0.0
